requirements.txt
asgiref
certifi
charset-normalizer
crispy-bootstrap5
Django
django-admin-list-filter-dropdown
django-crispy-forms
django-thaidate
idna
lxml
mysqlclient
python-docx
pytz
requests
sqlparse
typing_extensions
urllib3
bpython
